LayoutTests/media/picture-in-picture/picture-in-picture-interruption-expected.txt

This tests that video is in the correct media session state after ending a background interruption while pipped.

RUN(internals.settings.setAllowsPictureInPictureMediaPlayback(true))
RUN(internals.setMockVideoPresentationModeEnabled(true))
RUN(video.src = findMediaFile("video", "../content/test"))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EVENT(webkitpresentationmodechanged)
RUN(internals.beginMediaSessionInterruption("enteringbackground"))
EXPECTED (internals.mediaSessionState(video) == 'Playing') OK
RUN(internals.endMediaSessionInterruption("enteringbackground"))
EXPECTED (internals.mediaSessionState(video) == 'Playing') OK
END OF TEST
